Refactored style on Home to CSS-file and added media query to change background image on smaller screens
